Jogando main page pro lugar da swipe page
