+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:grupolaranja20212/pages/swipe_page.dart';
+import 'package:grupolaranja20212/pages/main_page.dart';
 import 'package:grupolaranja20212/pages/welcome_page.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
@@ -14,7 +14,7 @@ void main() async {
   Widget _defaultHome = const WelcomePage();
 
   if (FirebaseAuth.instance.currentUser != null) {
-    _defaultHome = const SwipePage();
+    _defaultHome = const MainPage();
   }
 
   // Run app!
