+++ lib/pages/auth_register_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:grupolaranja20212/pages/user_register_page.dart';
+import 'package:grupolaranja20212/pages/main_page.dart';
 import 'package:grupolaranja20212/view_models/register_view_model.dart';
 import 'package:provider/provider.dart';
 
@@ -27,7 +27,7 @@ class _AuthRegisterPage extends State<AuthRegisterPage> {
       isRegistered = await _registerVM.register(email, password);
       if (isRegistered) {
         Navigator.of(context).pushAndRemoveUntil(
-            MaterialPageRoute(builder: (context) => const UserRegisterPage()),
+            MaterialPageRoute(builder: (context) => const MainPage()),
             (Route<dynamic> route) => false);
       }
     }
