+++ lib/utils/app_navigator.dart
@@ -5,18 +5,13 @@ import 'package:grupolaranja20212/pages/auth_logout_page.dart';
 import 'package:grupolaranja20212/pages/auth_register_page.dart';
 import 'package:grupolaranja20212/pages/main_page.dart';
 import 'package:grupolaranja20212/pages/match_chat_page.dart';
-import 'package:grupolaranja20212/pages/matches_map_page.dart';
-import 'package:grupolaranja20212/pages/matches_page.dart';
 import 'package:grupolaranja20212/pages/profile_page.dart';
-import 'package:grupolaranja20212/pages/swipe_page.dart';
-import 'package:grupolaranja20212/pages/user_register_page.dart';
 import 'package:grupolaranja20212/provider/card_provider.dart';
 import 'package:grupolaranja20212/view_models/forgot_password_view_model.dart';
 import 'package:grupolaranja20212/view_models/login_view_model.dart';
 import 'package:grupolaranja20212/view_models/logout_view_model.dart';
 import 'package:grupolaranja20212/view_models/matches_view_model.dart';
 import 'package:grupolaranja20212/view_models/register_view_model.dart';
-import 'package:grupolaranja20212/view_models/swipe_view_model.dart';
 import 'package:provider/provider.dart';
 
 class AppNavigator {
@@ -50,26 +45,6 @@ class AppNavigator {
             fullscreenDialog: true));
   }
 
-  static Future navigateToMatchesPage(BuildContext context) async {
-    return await Navigator.push(
-        context,
-        MaterialPageRoute(
-            builder: (context) => ChangeNotifierProvider(
-                create: (context) => MatchesViewModel(),
-                child: const MatchesPage()),
-            fullscreenDialog: false));
-  }
-
-  static Future navigateToMatchesMapPage(BuildContext context) async {
-    return await Navigator.push(
-        context,
-        MaterialPageRoute(
-            builder: (context) => ChangeNotifierProvider(
-                create: (context) => MatchesViewModel(),
-                child: const MatchesMapPage()),
-            fullscreenDialog: false));
-  }
-
   static Future navigateToForgotPasswordPage(BuildContext context) async {
     return await Navigator.push(
         context,
@@ -77,16 +52,6 @@ class AppNavigator {
             builder: (context) => ChangeNotifierProvider(
                 create: (context) => ForgotPasswordViewModel(),
                 child: const AuthForgotPasswordPage()),
-            fullscreenDialog: false));
-  }
-
-  static Future navigateToUserRegisterPage(BuildContext context) async {
-    return await Navigator.push(
-        context,
-        MaterialPageRoute(
-            builder: (context) => ChangeNotifierProvider(
-                create: (context) => MatchesViewModel(),
-                child: const UserRegisterPage()),
             fullscreenDialog: false));
   }
 
@@ -116,16 +81,6 @@ class AppNavigator {
         MaterialPageRoute(
             builder: (context) => ChangeNotifierProvider(
                 create: (context) => CardProvider(), child: const MainPage()),
-            fullscreenDialog: false));
-  }
-
-  static Future navigateToSwipePage(BuildContext context) async {
-    return await Navigator.push(
-        context,
-        MaterialPageRoute(
-            builder: (context) => ChangeNotifierProvider(
-                create: (context) => SwipeViewModel(),
-                child: const SwipePage()),
             fullscreenDialog: false));
   }
 }
